tests: suppress normal integration output unless test failed/skipped
This commit modifies the output of Skeema's integration test suites, as well
as a few other tests. The purpose of this change is to make normal test output
less noisy and intimidating, and to make failing tests easier to diagnose.

Integration test output (both STDOUT and STDERR) is now buffered and
suppressed, and is only displayed if the test fails or is skipped. Any test
log annotations are displayed prior/above the full failing test output, rather
than being intermingled in that output.

Separately, two unit tests relating to error and panic handling now suppress
their output as well, since it is often confusing to new developers.

Finally, one test relating to `skeema diff --brief` now uses an in-memory
pipe for output capture, rather than a real file.