ci, go.mod: bump Go version to 1.21
Now that Golang 1.22 is available, the oldest Go version supported by the Go
authors is 1.21. Skeema follows the same logic for minimum Go version, so
this commit bumps go.mod and CI to specify 1.21.

Note that the release pipeline (powered by GoReleaser) always uses the latest
Go for building binary releases, so no change is necessary there.

No Golang 1.21 features are introduced in this commit, but subsequent commits
will do so, later today and further in the future. These 1.21 features are
most likely to be used soon:

* testing.Testing() for detecting test suite execution
* min, max builtins
* clear builtin
* log/slog package (eventually replacing logrus)
* slices package
* maps package