logger: disable word-wrap with narrow terminal
If STDERR is a terminal, and the terminal width is less than 80 characters,
automatic word-wrapping is now disabled. Otherwise, log messages become
difficult to read, since the first 28 characters of each line are reserved for
the log message header (timestamp and level), or a blank indent of same length
on subsequent wrapped lines.

Previously in this situation, wrapping would occur at 80 characters despite
the terminal being narrower than that, which was not correct behavior.