internal refactors prior to MariaDB 11.3 support
This commit includes two refactors, which are being made prior to MariaDB
11.3 support (which will come later in a separate commit).

The first refactor is a small adjustment to the interface which specifies
whether an ALTER TABLE clause is unsafe. Previously, this interface accepted
no args. Now, a StatementModifiers value can be supplied. This will initially
be used for supporting MariaDB 11.3's ability to safely convert inet4 columns
to inet6. With the previous interface, there was no way to specify the Flavor,
which is now needed to determine if this conversion is safe.

The bulk of this commit is a second refactor -- to various tengo.Flavor
methods, variables, and callsites -- primarily to avoid having to maintain an
ever-growing list of Flavor values.

Previously, we would introduce a new Flavor variable for every major.minor
release series; these were used for flavor-gating logic, as well as in tests.
When this approach was first adopted, only 1 to 3 new major.minor release
series were added per year. But now that both MariaDB and MySQL (and variant
Percona Server) each release a new minor version every 3 months, this setup is
no longer ideal, as we already have 31 such values defined and the list is
growing by 12 values per year.

To solve this, new flavor comparison functions have been introduced, which no
longer rely on the pre-baked Flavor values. For example, code previously like
this:
    if flavor.Min(tengo.FlavorMySQL57) { ... }
now becomes this:
    if flavor.MinMySQL(5, 7) { ... }

Meanwhile, in test logic, flavors are now often expressed as strings which are
then passed to ParseFlavor. This is less efficient than the old pre-baked
values, but the impact to test execution time should still be negligible.