flavor: improve handling of too-new db server version
Previously, attempting to use a release series that is newer than this version
of Skeema would result in a fatal error in workspace=docker, and no logging
whatsoever with workspace=temp-schema.

This commit changes this behavior, so that this situation logs a non-fatal
warning, regardless of the workspace option. There are a few motivations for
this adjustment:

* Most recent release series, for both MySQL and MariaDB, have only involved
  minor impacts to schema management. As long as a Docker image is available
  for a new DB server release series, using workspace=docker may function
  fine, especially if the Skeema release isn't too terribly old.

* Although we generally update the Skeema source code to handle new DB server
  release series within 1-2 weeks of their GA release, sometimes some time may
  pass until the next Skeema binary release, especially if the changes in the
  new DB server release series are minor. No sense in blocking
  workspace=docker from working in this situation.

* Some companies pin their Skeema release and then forget to upgrade for a
  long time, sometimes years. Having a warning for too-new DB server is useful
  as a reminder, and may also reduce support burden.

This commit includes a few other minor adjustments to related logging message
text, as well as some code refactors. In particular, the location for tracking
the latest-supported MySQL and MariaDB server versions is now found at the top
of internal/tengo/flavor.go, rather than being part of the logic of a method.
This means that when a new DB server release series comes out, if its
release notes don't indicate any changes relevant to Skeema's logic, all that
needs to happen now is the adjustment of a variable at the top of this file.