workspace=docker: update Percona Server arm64 latest 8.0 image
When using workspace=docker with a "percona:8.0" flavor on arm64, Skeema's
Docker image selection logic must supply a specific patch release number.
This is necessary because the Percona Server image tags for "latest" or "8.0"
don't include arm64 builds on DockerHub. To get an arm64 image we must use a
tag of the form "8.0.x-aarch64" with a specific value for x.

Previously we used 8.0.35 for this, since that was the latest available at
that time. This commit adjusts the logic to use 8.0.36 now that it is on
DockerHub. This patch number is now stored in a string constant.