tests: fix two integration tests for Percona Server on arm64
When choosing a Percona Server 8.x image on arm64, special logic is needed
for determining the corresponding Docker image name on DockerHub. This logic
was recently introduced for workspace=docker, but was previously missing for
two integration tests (TestLowerCaseTableNames1 and TestTempSchemaBinlog)
which need to create special-case containerized databases.

This commit fixes these two integration tests, and also exports a previously
unexported method from the workspace package for re-use here.