modules: upgrade github.com/go-sql-driver/mysql to v1.8.1
This commit updates dependency github.com/go-sql-driver/mysql (the Golang
MySQL driver) to v1.8.1, which was released in late March as a small patch on
top of v1.8.0 from early March.

Previously we used v1.7.1 from a year ago, so this is a somewhat large driver
change. This is intentionally being done right *after* a Skeema release, so
that it has more time being tested before the next Skeema release.

That said, all integration tests pass, so no initial concerns with this
upgrade for Skeema Community Edition.

Separately, some minor changes are necessary in Skeema Premium Edition
regarding seed views for SET and ENUM types, and a couple changes to
integration tests to use some new 1.8 driver improvements.