diff/push: detect index renaming and emit RENAME KEY
This commit enhances Skeema's diff logic to automatically detect when an index
is being renamed, and emit ALTER TABLE ... RENAME KEY accordingly if the
database server supports this syntax.

RENAME KEY allows an index to be renamed without having to DROP and re-ADD it.
It is present in MySQL 5.7+ and MariaDB 10.5+. In older database servers,
Skeema will fall back to using DROP and re-ADD, as before.

Index rename detection normally depends on the index's definition having no
other changes to the index whatsoever. However, as a special case, adjustments
to an index's COMMENT clause are ignored for rename purposes IF Skeema's
lax-comments option is enabled. Aside from that specific situation, changes to
the index's definition will force Skeema to fall back to DROP and re-ADD.

Note that index renames are not compatible with changing an index's visibility
(INVISIBLE / VISIBLE in MySQL, or IGNORED / NOT IGNORED in MariaDB) in the
same ALTER TABLE, in any known database server version, even though both
operations are normally in-place. In this situation, Skeema currently falls
back to DROP and re-ADD.