lint-reserved-word: handle one missing special-case word
The parser used by MySQL (and MariaDB) treats "_filename" as a reserved word,
even though it is not listed in information_schema.keywords, nor is it present
on the normal list of reserved words in the manual. The MySQL manual mentions
this separately as a special case. The MariaDB manual does not appear to
mention it, but its behavior matches MySQL for this reserved word.

This commit updates lint-reserved-word to treat "_filename" as a reserved
word in all flavors.