internal refactor: SQL error number handling
This commit refactors code relating to detection of specific server error
numbers. Previously, this logic relied on constants in an external package
github.com/VividCortex/mysqlerr, but that package is obsolete: it is no
longer being updated with new errors from recent server releases.

In this commit, we now track only the relevant error codes internally, and
expose new helper functions which are used by the appropriate callsites.

This refactor was motivated by an overall desire to avoid unnecessary
external module dependencies, which was made more urgent due to MySQL 8.4's
introduction of new error codes (to be handled in a subsequent commit).

This commit also removes tengo.IsAccessError(), which was dead code.