linter: ignore incorrect flavor overrides
Some linter checks rely on the flavor value in their logic or annotation text.
Previously, this value was populated based on the configured value of the
"flavor" option, from either the directory .skeema file or a command-line
override. However, the configured value does not necessarily match the
workspace database instance, especially when using the default of
workspace=temp-schema. For example, if you upgrade your database server, the
.skeema file's flavor value is only updated automatically if you run `skeema
pull`.

This commit solves the problem by using the actual flavor value from the
workspace database instance. This ensures the introspected schema and flavor
value are aligned, and has the added benefit of making the specific patch
(point release) version number available to linter logic as well.