lint-fk-parent: adjust annotation text
When MySQL 8.4 was first released, its documentation did not mention the
foreign key changes in this release; see discussion on MySQL bug 114838.
However, Oracle has now updated their release notes and manual to cover this
change, which points to bug 97836 as a primary motivator: if the parent side
of a foreign key constraint lacks a UNIQUE KEY on the exact referenced
columns, this can break group replication when multi-threaded applier is in
use.

Now that the full motivation for this FK change is known, this commit updates
the annotation message for lint-fk-parent accordingly.

More background in our blog post, which has now been updated as well:
https://www.skeema.io/blog/2024/05/14/mysql84-surprises/#foreign-key-restrictions