internal workspace tweak: use result struct for instance metadata
In recent commit c151901, whenever a workspace schema is introspected, the
underlying database instance's flavor value was added to the returned
information. In that commit, the flavor was returned as a separate return
value.

The shortcoming to that approach is awkwardness in adding further metadata
fields to the introspection result. So this commit now refactors that logic
slightly, in order to use a new intermediate result struct. In addition, the
returned result now includes the workspace session's sql_mode value, alongside
the introspected schema and the instance flavor.

The sql_mode value is not used by any Community Edition features yet, but it
may be useful in the future, and it will immediately be used in a new Premium
Edition feature under development.