diff/push: improve RENAME KEY when also changing visibility
Support for ALTER TABLE ... RENAME KEY was recently added in commit 478ceaf.
However, that commit didn't handle one particular edge case in an ideal
fashion: attempting to rename an index while also changing the visibility of
the index. MySQL and MariaDB prohibit performing both operations on the same
index in the same ALTER TABLE, so this was originally handled by always
falling back to DROP-and-re-ADD in this specific situation.

Today's commit now improves that logic to emit two separate ALTER TABLEs in
this situation: the first with the RENAME KEY (and any other alter clauses),
and the second with just the ALTER INDEX to change visibility. Even though
this entails a second ALTER TABLE, when executing DDL directly (no OSC
wrapper) these two ALTER TABLEs will perform better than fully rebuilding an
index if it is of non-trivial size.

In some extremely rare edge cases, the first ALTER TABLE must still DROP-and-
re-ADD the index. In these situations, the second ALTER TABLE is suppressed,
since the re-ADD will already contain the desired visibility:

* Renaming the index and also changing its COMMENT clause, without using
  the --lax-comments option
* Renaming the index and also re-ordering a prior index, when using
  the --exact-match option

All of the situations handled by this commit are quite rare, but are
nonetheless being addressed here for sake of completeness of the RENAME KEY
handling ahead of its initial inclusion in an upcoming release.