ci: ignore errors from Coveralls
This commit adds "continue-on-error: true" to the Coveralls step, as Coveralls
is consistently returning 504 errors today, blocking Skeema v1.12.0's release.

This was previously used to avoid an issue in Sep 2021 in fdbe495, but was
reverted a month later in 9a55af6. We may need to leave it in place
permanently this time and/or consider moving to a different provider.