tests: prevent flaky TestInstanceProcessList failure
In very rare cases, apparently the server can return -1 for the TIME field
of a processlist entry. This causes tengo.Instance.ProcessList() to return an
error, since the time field was being read into an unsigned int. This then
causes an integration test failure for TestInstanceProcessList in the tengo
subpackage.

This commit fixes the problem by switching to a signed int scan destination.

The tengo.Instance.ProcessList method is only called by this one integration
test, and not used by any Skeema code paths or behaviors, since it is only
present for ad hoc debugging of containerized test database misbehavior.