ci: pin MariaDB 10.11 test version to 10.11.8 for now
Integration tests are failing on MariaDB 10.11.9 (released last week) in two
places. One spot relates to foreign_key_checks, and the other relates to
information_schema representation of compressed columns. These test failures
are unexpected, as nothing in the MariaDB release notes or changelogs suggests
an intentional behavior adjustment with either of these server features. More
investigation is required.

For now, CI is being pinned to the previous quarter's point release to avoid
the test failures.