ci, go.mod: bump Go version to 1.22
Now that Golang 1.23 is available, the oldest Go version supported by the Go
authors is 1.22. Skeema follows the same logic for minimum Go version, so
this commit bumps go.mod and CI to specify 1.22.

Note that the release pipeline (powered by GoReleaser) always uses the latest
Go for building binary releases, so no change is necessary there.

No Golang 1.22 features are introduced in this commit, but subsequent commits
may do so in the future. These 1.22 changes are most likely to be used soon:

* database/sql generic Null type
* slices.Concat
* cmp.Or
* for loop variable per-iteration scope fix
* for loop range over integers