Percona Server 8.4 support (fix for workspace=docker on arm64)
This commit adds support for Percona Server for MySQL 8.4, released yesterday.
In particular, workspace=docker needs special handling for Percona 8.4.x on
systems with ARM CPUs, due to how Percona Server ARM images are tagged on
DockerHub. Aside from this, Percona Server 8.4 was already supported by Skeema
as-is when using workspace=temp-schema, or when using workspace=docker on an
Intel/AMD system.

This commit also bumps the patch release number for flavor=percona:8.0 on ARM
with workspace=docker to use 8.0.37, which was released earlier this month.

We will file a bug report with Percona soon to see if their ARM image tagging
can be corrected, which will remove the need for tracking the specific patch
releases.