workspace=docker: improve image selection for Percona on arm64
In most cases, workspace=docker selects a Docker image based on the value of
the flavor option, as-is. And the flavor option typically omits the specific
patch version -- `skeema init` auto-populates it with just the major.minor
version, since the patch release number is rarely necessary, as the Docker
image tag for major.minor will normally point to the latest patch.

However, Percona Server releases on DockerHub only have major.minor tags for
AMD/Intel builds and not ARM builds. We must supply a patch release somehow
when running on ARM in order to select an image. Previously, we hard-coded the
latest patch number at the time of each Skeema release, and used that.

This commit improves the logic to examine the flavor of the "real" target DB
server, similar to how we examine its sql_mode as well. If the flavor option
is a Percona Server major.minor value e.g. flavor=percona:8.0, and the target
DB server matches that, we now plug in the DB server's real patch version to
obtain the correct Docker image on ARM.

In some situations, this new logic cannot be applied, and it is skipped
accordingly. For example, some users override --flavor on the command-line in
order to test cross-flavor compatibility/linting of their schema definitions.
We don't copy the real server's patch value if its flavor family is not
aligned with the requested flavor. In that situation we still fall back to the
hard-coded "latest patch at Skeema release time" values for Percona Server on
ARM.

This logic is effectively a no-op for Innovation releases 8.1, 8.2, 8.3 since
those only ever have a patch of 0. And this won't be relevant for Innovation
releases 9.0-9.6, since Percona Server is skipping non-LTS releases moving
forwards. (That also means we don't need a special-case for handling 9.0.1,
which was released by Oracle as an emergency fix, even though normally
Innovation releases stop after their initial .0 release.)

This commit does not affect non-ARM image selection, nor non-Percona Server
image selection.