MariaDB 11.6 support
This commit marks MariaDB 11.6 as a supported DB flavor, now that its GA
release occurred on 21 Nov 2024. This is a rolling (short term support)
MariaDB server release.

This is just a small bookkeeping commit. No functionality changes were
necessary for MariaDB 11.6 support in Skeema, as all integration tests already
pass as-is.

Two notes for users:

* MariaDB 11.6 changes the server default character set to utf8mb4. This
  default change typically does not directly affect Skeema's management of
  existing databases. Each schema normally has its own default set in a
  .skeema file using option default-character-set, which overrides the server-
  level default.

* MariaDB 11.6 introduces a new authentication plugin called PARSEC, which is
  not supported yet by Skeema. This will require waiting for support in the
  Golang database driver.