init, add-environment: persist TLS options to .skeema file
This commit adjusts `skeema init` and `skeema add-environment` so that TLS-
related options from the command-line are automatically persisted to the
.skeema option file alongside the host name. This is more consistent with how
other connectivity-related options are handled, such as user and
connect-options.

In Skeema Community Edition, this change only affects the ssl-mode option.
Meanwhile in Skeema Premium Edition it also affects the enhanced TLS options
(ssl-cert, ssl-key, ssl-ca, etc).

This commit also refactors the underlying logic so that the list of persisted
options is not duplicated between cmd_init.go and cmd_add_environment.go.