MySQL 9.2 support
This commit marks MySQL 9.2 as a supported DB flavor. MySQL 9.2 is a short-
term-support Innovation server version, released yesterday (21 Jan 2025).

This is a small bookkeeping commit which adds 1 new reserved word (LIBRARY)
which is used by linter check lint-reserved-word. No other changes in
functionality were necessary for MySQL 9.2 support in Skeema, and all
integration tests already pass as-is.

Note: MySQL 9.2 includes some privilege changes related to spatial reference
systems, but Skeema does not support management of CREATE SPATIAL REFERENCE
SYSTEM statements yet at this time. We can evaluate adding support for that
if/when there is community interest in such a feature; please open an issue or
GitHub discussion if this would be useful for your company.

Unrelated to MySQL 9.2 but affecting the same file, this commit also updates
the tracking of the latest Percona Server patch numbers for 8.0.x and 8.4.x.
We track these values for use on ARM systems, since there is no "8.0" or
"8.4" tags on DockerHub for arm64.