timeouts: ensure lock wait timeouts occur before readTimeout
By default, Skeema uses a readTimeout (query timeout) of 20 seconds for many
of its queries, to prevent the appearance of a stall/freeze if a server is
non-responsive or overloaded. However, this readTimeout makes troubleshooting
more difficult in situations involving server metadata lock waits
(lock_wait_timeout) or InnoDB lock waits (innodb_lock_wait_timeout), as both
of these server variables have defaults higher than 20 seconds. As a result,
the `skeema` invocation will error with an i/o timeout (hiding the true source
of the wait) and meanwhile the queries might remain running on the server
until hitting the relevant server timeout.

This commit improves the situation by ensuring that whenever a readTimeout is
applied, the session will set lock_wait_timeout and innodb_lock_wait_timeout
to be lower than the readTimeout. This is done in a safe manner, such that
if the server *already* uses an even lower wait timeout value, the lower value
wins.

Additionally, this commit ensures that readTimeout cannot be set below 2
seconds, as a practical matter to ensure 1-second wait timeouts will always be
under the readTimeout. Attempting to set a lower readTimeout inside Skeema's
connect-options option will now result in readTimeout=2s being applied. (This
is not expected to affect many real users, as such a low readTimeout is rarely
practical anyway.)