diff/push: require --allow-unsafe if functional index collation change
Since Skeema v1.11, changing the collation of a column used in a uniqueness
constraint requires --allow-unsafe. This is because collation changes affect
the equality operation in the uniqueness constraint and can result in the
ALTER failing with a duplicate key error; see commit 905ee14.

However, the original implementation did not flag such changes when the unique
index is a functional index, which are supported in MySQL 8+. This commit
improves the underlying method used by the unsafe change detection to catch
this additional edge case.

Note that MariaDB does not directly support functional indexes, so this change
does not affect MariaDB. (The equivalent in MariaDB is to create a virtual
column and an index on that column, and Skeema's existing logic would already
handle that.)

This commit also includes a couple new table utility methods, which are not
used yet but may be helpful in the future.