tests: handle nonstandard MariaDB character_set_collations overrides
MariaDB 11.2+ provides the character_set_collations server variable, which
can be used to override the normal default collation associated with one or
more character sets. In stock MariaDB, this variable is blank in 11.2 thru
11.4, and then in 11.5+ it sets uca1400 collation defaults for 5 charsets.

However, some Linux distributions override this variable out-of-the-box in
nonstandard ways. In particular, Debian overrides it in MariaDB 11.4, and
Docker images for MariaDB inherit that override. This breaks some Skeema
integration tests when using SKEEMA_TEST_IMAGES=mariadb:11.4.

This commit solves the problem by making a few tests aware of that server
variable, by using a new helper function which dynamically queries the
server's character_set_collations as needed.