flavors: deprecate MySQL 5.5 and MariaDB 10.1
This commit marks flavors MySQL 5.5 and MariaDB 10.1 as deprecated by Skeema.
Any use of these flavors (either as a target database server, or when using
workspace=docker) now causes a warning message to be logged.

Aside from the warning, these flavors are still fully supported by Skeema v1.
However, support for these flavors is planned to be dropped in Skeema v2 due
to their extreme age, as described in this blog post:
https://www.skeema.io/blog/2025/05/12/skeema-v2-roadmap-rfc/