internal: refactor identifier parsing logic
This commit refactors the logic for parsing object names out of CREATE
statements, in order to create a reusable method for parsing identifier
clauses in general. The revised logic also accounts for situations with
multiple qualifiers (e.g. schemaname.tablename.colname) in case this is
useful in the future.

No changes to user-facing behavior of any Skeema subcommands.