modules: upgrade github.com/go-sql-driver/mysql to v1.9.3
This commit updates dependency github.com/go-sql-driver/mysql (the Golang
MySQL driver) to v1.9.3, which was released last Friday. This contains a fix
for go-sql-driver/mysql#1718 which was a blocker for our use of compressed
connections. An upcoming Skeema commit will make use of this compression
functionality in queries with heavy result-sets (e.g. introspecting stored
procedure bodies) whenever the observable latency to the server is above a
threshold.

Thank you to the go-sql-driver/mysql maintainers for a speedy fix and release!