push: deprecate concurrent-instances, replace with concurrent-servers
This commit renames option concurrent-instances to concurrent-servers. The old
concurrent-instances name will still function as an alias in Skeema v1, but is
now deprecated, and planned for removal in Skeema v2.

This change is consistent with other user-facing text changes made last year
in commit 5f7598e.

After this change, only the "-- instance: ..." separator comment in STDOUT
output of `skeema diff` and `skeema push` still uses the old terminology. This
will likely change in Skeema v2, but definitely not earlier, to avoid breaking
existing users who parse this text.