MySQL 9.4 support
This commit marks MySQL 9.4 as a supported DB flavor. MySQL 9.4 is a short-
term-support Innovation server version, released yesterday (22 July 2025).

The only necessary change here relates to integration testing: the `mysql`
command-line client now has a --commands option, which defaults to true in
8.4 LTS, but false in 9.4+. This means that by default, integration test .sql
files with client command usage would otherwise start failing in 9.4.

One new oddity is that with --commands=false, the USE command is processed
server-side instead of client-side, which means it must be followed by a
delimiter in order to function. In contrast, with client-side command
processing, a newline is sufficient without any delimiter; several of our
integration test fixtures do indeed use that style. So corresponding logic
now ensures --commands is enabled for 9.4+.

Aside from that, no changes in functionality were necessary in Skeema, as
MySQL 9.4 does not contain any features that affect schema management.

Users of workspace=docker, please note that DockerHub currently only has amd64
images of MySQL 9.4 and not arm64. Meanwhile the "mysql:9" tag now points to
9.4 for amd64 but still 9.3 for arm64, which is quite strange. We will file an
issue with the MySQL Docker maintainers soon if this remains the case.