stored program introspection: strip superfluous trailing semicolon
When a stored program (proc, func, trigger, event) is created by a client with
multi-statement support enabled along with a non-default DELIMITER, and a
trailing semicolon was present after the end of the body, the server retains
this semicolon in the body in output of SHOW CREATE. However, any other method
of creating the stored program will strip it.

Since Skeema does not create stored programs using multi-statements, this can
cause spurious diffs which persist across multiple `skeema push` invocations.
To solve this, such trailing semicolons are now stripped at introspection
time.