sql files: auto-ignore CREATE TABLE with wrong delimiter
If a .sql file contains a DELIMITER command which sets a non-default (non-
semicolon) delimiter, and then contains a CREATE TABLE statement ending in a
semicolon, that CREATE TABLE is now automatically ignored, along with any
other statements prior to the correct DELIMITER string. This form of typo
could otherwise cause unexpected results when executed in a workspace with
multi-statements enabled.

When a CREATE TABLE is ignored in this fashion, a warning is logged in the
same fashion as other auto-ignored types of CREATE TABLE (e.g. system-
versioned tables, CREATE ... SELECT statements).

Although Skeema does not yet use multi-statements as of this commit, a
subsequent commit will optionally make use of them to reduce workspace round-
trips in high-latency scenarios. This functionality will only ever be used
for tables, since tables do not permit compound statements (BEGIN...END
blocks) and therefore have no valid need for non-default delimiters in the
first place.