tests: refine workaround for MariaDB Q3 2024 FK bug
MariaDB 10.11+ point releases from a year ago contained a bug which affected
foreign keys. We reported this as MDEV-34756, and MariaDB engineers promptly
fixed the bug in the following quarter's releases. At the time, to avoid
failing Skeema integration tests, a workaround was applied for all of MariaDB
10.11.9+.

Now that these affected point releases are a year old, we can scale back this
workaround, since we typically run integration suites against the latest point
release in each version series and the bug was fixed for 10.11, 11.2, 11.4,
and 11.6+. However, the workaround cannot be removed entirely, since two of
the affected version series (11.1 and 11.5) went EOL with the point release
that had the FK bug.