temp-schema: fix create-chunking when only using temp-schema-threads
The new create-chunking optimization was intended to be disabled for users who
are setting the old temp-schema-threads option and not the newer
temp-schema-mode option. However, a code refactor before merge inadvertently
caused the default chunking behavior of temp-schema-mode=regular to be applied
in this situation as well. This commit fixes the logic to force chunk size to
1 in this case.