chore: tweak GoReleaser changelog exclusion prefixes
This commit adjusts the commit summary prefixes that are skipped in GoReleaser
changelog generation. Although the auto-generated changelog isn't used
directly (in favor of human-written release notes), it is still useful output
when cutting a new release, to confirm that all relevant commits are included
in the final release notes.