MariaDB 12.0 support
This commit marks MariaDB 12.0 as a supported DB flavor. This is MariaDB's
Q3 rolling release, which became GA in August 2025.

One change in MariaDB 12.0 affects schema management:

* MDEV-10164: Triggers may now be defined to fire on a combination of
  multiple events (INSERT/UPDATE/DELETE) separated by " OR ", instead of just
  a single event.

In Skeema Community Edition, this is just a bookkeeping commit to bump the
max supported MariaDB version. The trigger change does not affect Skeema
Community Edition (which does not support triggers), and all integration tests
already pass against MariaDB 12.0 without any functionality changes.

In Skeema Premium Edition, multiple-event triggers are now supported. However
please note that when a table has multiple triggers affecting a timing and
event pair, but at least one of those triggers is a multiple-event one, it is
not possible to define a user-controllable ordering between the single-event
and multi-event triggers due to a server-side limitation. See MDEV-37711 for
more information.