generator: error or warn on mismatched Skeema versions
This commit adds checks against the generator option value, to detect
situations where the directory structure was generated using a newer version
of Skeema (and/or more enhanced edition) relative to the current binary being
executed. These checks are run at the start of any command that operates on
existing *.sql files.

For background, when `skeema init` creates a directory structure, the top-
level .skeema file will contain a generator line, reflecting the version and
edition of Skeema used, for example "skeema:1.13.1-community". Future
invocations of `skeema pull` update this value whenever appropriate. This
value is used in gating some behavior for compatibility purposes, especially
when first upgrading from FOSS to paid editions of Skeema.

This commit expands the use of the generator value for purposes of detecting
version or edition mismatches:

* If the generator was a newer major version of Skeema (e.g. generator is
  Skeema v2 but you are running Skeema v1), an error is logged and the program
  aborts. This is necessary since new major versions will contain backwards-
  incompatible behaviors, as by definition that's why the major version number
  was incremented.

* If the generator was the same major version but a newer minor version, a
  non-fatal warning is logged. This is helpful because the configuration may
  rely on options or features that don't exist in the older minor version
  binary.

* If the generator was a paid edition of Skeema, but you are running a FOSS
  binary, a non-fatal warning is logged. This is helpful because the config
  may rely on options or features that don't exist in the FOSS binary.

* If the generator indicates it was something other than Skeema, a non-fatal
  warning is logged. This could be helpful in the future if other third-party
  tools wish to interoperate with Skeema's directory format.