lint-reserved-word: handle EXTERNAL (from MySQL 9.4)
This commit fixes an omission in lint-reserved-word: "external" is a reserved
word in MySQL 9.4+.

This was accidentally missed in our MySQL 9.4 compatibility commit (b231f76)
because its row in information_schema.keywords erroneously lists it as non-
reserved. Furthermore, nothing in the MySQL 9.4 release notes directly
mentions this keyword, nor any feature relating to it. However, the MySQL
manual does correctly list it as a reserved word.