docker: fix recent race condition on slow machines
Last month's connection-reuse refactor 49b79d1 inadvertently introduced a
race condition when interacting with a database server that is being
initialized for the first time. In this situation, a connection can
succeed but then get dropped immediately as the server restarts during its
initialization process; the refactored code could then cache a bad conn.

In practical terms, this only affects Skeema-managed Docker instances, as that
is the only realistic situation in which a connection is launched against a
server that is being newly initialized. Furthermore the race condition only
seems to occur on slower machines, and possibly only with older flavor images.
In any case, this commit catches this scenario, so the "is the Docker
container reachable yet" code path won't return a false positive.

This commit also has the side-effect of making the server variable
introspection logic mandatory, whereas previously failures in that query were
intentionally ignored. This is not expected to have any user-facing impact, as
there are no known supported server flavors where the queried global variables
do not exist.